Add and use concern to handle duplication of agenda items as well
